PR 4: Tool Interface Abstraction
- BaseToolset — abstract interface with act()/action()/exclude_action()
- BrowserToolset — 24 built-in browser actions (extracted from Tools)
- CustomToolset — lightweight class for user-defined actions
- Tools — composition pattern, contains .browser + .custom, shares registry
- Agent depends on BaseToolset type hint, accepts any toolset subclass

Zero behavior change — all existing APIs and imports preserved.
