+++ examples/custom-functions/file_upload.py
@@ -18,10 +18,11 @@ from dotenv import load_dotenv
 load_dotenv()
 
 from browser_use import ChatOpenAI
-from browser_use.agent.service import Agent, Tools
+from browser_use.agent.service import Agent
 from browser_use.agent.views import ActionResult
 from browser_use.browser import BrowserSession
 from browser_use.browser.events import UploadFileEvent
+from browser_use.tools.service import Tools
 
 logger = logging.getLogger(__name__)
